telegram_bot/requirements.txt
python-telegram-bot==20.7
pandas>=1.3.0
openpyxl>=3.0.0